k8s-simulator/backend/requirements.txt
# [advice from AI] K8S 시뮬레이터 백엔드 의존성 패키지들
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
pyyaml==6.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
httpx==0.25.2
websockets==12.0
prometheus-client==0.19.0
structlog==23.2.0
